+++ CHANGELOG.md
@@ -26,6 +26,7 @@ metric events are swallowed for the time being.
 - Add sensuctl extension command.
 - Add extensions to pipelined.
 - Added more tests surrounding the sensu-agent's statsd server and udp port.
+- Add default user with username "sensu" with global, read-only permissions.
 
 ### Changed
 - Changed the maximum number of open file descriptors on a system to from 1024